fix: split migration script
